src/test/java/com/oranshuster/webcrawlerverifier/bots/SottopopBotDataTest.java
package com.oranshuster.webcrawlerverifier.bots;

import com.google.common.collect.ImmutableList;
import nl.basjes.parse.useragent.UserAgent;
import nl.basjes.parse.useragent.UserAgentAnalyzer;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.Test;

import java.util.Collections;

import static org.testng.Assert.*;

public class SottopopBotDataTest {
    private UserAgentAnalyzer uaa;

    @BeforeClass
    public void beforeClass() {
        uaa = UserAgentAnalyzer.newBuilder()
                .hideMatcherLoadStats()
                .withCache(10000)
                .build();
    }

    @Test
    public void testGetIdentifier() throws Exception {
        assertEquals(SottopopBotData.getInstance().getIdentifier(), "SOTTOPOPBOT");
    }

    @Test
    public void testGetUserAgentChecker() throws Exception {
        UserAgent ua = uaa.parse("Sottopop/1.0(+https://upcontent.com/robots)");
        assertTrue(SottopopBotData.getInstance().getUserAgentChecker().test(ua));

        ua = uaa.parse("Sottopop(+https://upcontent.com/robots)");
        assertTrue(SottopopBotData.getInstance().getUserAgentChecker().test(ua));

        ua = uaa.parse("Good browser");
        assertFalse(SottopopBotData.getInstance().getUserAgentChecker().test(ua));
    }

    @Test
    public void testGetIps() throws Exception {
        assertTrue(SottopopBotData.getInstance().getIps().isEmpty());
    }

    @Test
    public void testGetHostnames() throws Exception {
        assertEquals(SottopopBotData.getInstance().getHostnames(), ImmutableList.of("upcontent.com"));
    }

}